+++ cmp/grid/impl/ColumnGroupHeader.js
@@ -5,7 +5,7 @@
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 import {div, span} from '@xh/hoist/cmp/layout';
-import {hoistCmp, HoistModel, useLocalModel} from '@xh/hoist/core';
+import {hoistCmp, HoistModel, creates} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
 import {bindable, makeObservable} from '@xh/hoist/mobx';
 import classNames from 'classnames';
@@ -19,54 +19,60 @@ import {isFunction} from 'lodash';
 export const columnGroupHeader = hoistCmp.factory({
     displayName: 'ColumnGroupHeader',
     className: 'xh-grid-group-header',
-    model: false,
+    model: creates(() => ColumnGroupHeaderModel),
 
-    render(props) {
-        const impl = useLocalModel(() => new LocalModel(props)),
-            expandIcon = () => {
-                if (!impl.isExpandable) return null;
+    render({model, className, xhColumnGroup, gridModel, setExpanded, displayName}) {
+        const {isExpandable, isExpanded} = model;
+        const expandIcon = () => {
+            if (!isExpandable) return null;
 
-                const icon = impl.isExpanded ? Icon.chevronLeft({size: 'sm'}) : Icon.chevronRight({size: 'sm'});
-                return div(icon);
-            };
+            const icon = isExpanded ? Icon.chevronLeft({size: 'sm'}) : Icon.chevronRight({size: 'sm'});
+            return div(icon);
+        };
 
         // We will only have an xhColumnGroup if this was a configured column group. ag-Grid will
         // auto-create column groups to ensure the headers are balanced for columns which are not
         // in a group being rendered next to columns which are in groups.
-        let headerName = props.displayName;
-        const {xhColumnGroup, gridModel} = props;
+        let headerName = displayName;
         if (xhColumnGroup && isFunction(xhColumnGroup.headerName)) {
             headerName = xhColumnGroup.headerName({columnGroup: xhColumnGroup, gridModel});
         }
 
         return div({
-            className: classNames(props.className, impl.isExpandable ? 'xh-grid-group-header--expandable' : null),
+            className: classNames(className, isExpandable ? 'xh-grid-group-header--expandable' : null),
             items: [
                 span(headerName),
                 expandIcon()
             ],
             onClick: (e) => {
-                if (!impl.isExpandable) return;
+                if (!isExpandable) return;
 
                 e.stopPropagation();
-                props.setExpanded(!impl.isExpanded);
+                setExpanded(!isExpanded);
             }
         });
     }
 });
 
-class LocalModel extends HoistModel {
-    /** @member {ColumnGroup} */
-    agColumnGroup;
+class ColumnGroupHeaderModel extends HoistModel {
 
     @bindable isExpanded = true;
 
-    get isExpandable() {return this.agColumnGroup.isExpandable()}
+    get isExpandable() {
+        return this.agColumnGroup.isExpandable();
+    }
 
-    constructor({columnGroup: agColumnGroup}) {
+    /** @member {ColumnGroup} */
+    get agColumnGroup() {
+        return this.componentProps.columnGroup.providedColumnGroup;
+    }
+
+    constructor() {
         super();
         makeObservable(this);
-        this.agColumnGroup = agColumnGroup.providedColumnGroup;
+    }
+
+    onLinked() {
         this.syncIsExpanded();
         this.agColumnGroup.addEventListener('expandedChanged', this.syncIsExpanded);
     }