+++ data/cube/aggregate/Aggregator.js
@@ -7,12 +7,25 @@
 
 export class Aggregator {
 
+
+    /**
+     * Does this aggregator depend only on leaf nodes contained by the node being aggregated?
+     *
+     * By default this property returns true, indicating to the cube that if all children of a
+     * given node are the same, the node does not need to be re-aggregated.  Aggregators that
+     * depend on global values (e.g. % of total) should return false for this value.
+     */
+    get dependsOnChildrenOnly() {
+        return true;
+    }
+
     /**
      * Aggregate values
      * @param {Object[]} rows - current rows in aggregation.  Will never be empty.
      * @param {string} fieldName - name of field to perform the aggregation on.
+     * @param {AggregationContext} context - current aggregation context
      */
-    aggregate(rows, fieldName) {}
+    aggregate(rows, fieldName, context) {}
 
 
     /**
@@ -23,10 +36,11 @@ export class Aggregator {
      * @param {RowUpdate} update - update that occurred to child of this aggregation.
      * Sub-classes may use this method to provide efficient implementations for dynamic changes
      * to an aggregation.  The default implementation will simply re-aggregate.
+     * @param {AggregationContext} context - current aggregation context
      * @returns {*} - new aggregate value
      */
-    replace(rows, currVal, update) {
-        return this.aggregate(rows, update.field.name);
+    replace(rows, currVal, update, context) {
+        return this.aggregate(rows, update.field.name, context);
     }