+++ mobile/cmp/input/TextInput.js
@@ -70,7 +70,7 @@ TextInput.hasLayoutSupport = true;
 class Model extends HoistInputModel {
 
     get commitOnChange() {
-        return withDefault(this.props.commitOnChange, false);
+        return withDefault(this.componentProps.commitOnChange, false);
     }
 
     onChange = (ev) => {
@@ -80,13 +80,12 @@ class Model extends HoistInputModel {
     };
 
     onKeyDown = (ev) => {
-        const {onKeyDown} = this.props;
         if (ev.key === 'Enter') this.doCommit();
-        if (onKeyDown) onKeyDown(ev);
+        this.componentProps.onKeyDown?.(ev);
     };
 
     onFocus = (ev) => {
-        if (this.props.selectOnFocus && ev.target && ev.target.select) {
+        if (this.componentProps.selectOnFocus && ev.target && ev.target.select) {
             ev.target.select();
         }
         this.noteFocused();